+++ scout/server/blueprints/variants/controllers.py
@@ -51,6 +51,8 @@ from scout.server.utils import (
 from .forms import FILTERSFORMCLASS, CancerSvFiltersForm, FusionFiltersForm, SvFiltersForm
 from .utils import update_case_panels
 
+NUM = re.compile(r"\d+")
+
 LOG = logging.getLogger(__name__)
 
 
@@ -938,6 +940,10 @@ def parse_variant(
     if not "end_chrom" in variant_obj:
         variant_obj["end_chrom"] = variant_obj["chromosome"]
 
+    # common motif count for STR variants
+
+    variant_obj["str_mc"] = get_str_mc(variant_obj)
+
     # variant level links shown on variants page
     variant_obj["cosmic_links"] = cosmic_links(variant_obj)
     variant_obj["str_source_link"] = str_source_link(variant_obj)
@@ -957,6 +963,31 @@ def parse_variant(
     ]
 
     return variant_obj
+
+
+def get_str_mc(variant_obj: dict) -> Optional[int]:
+    """Return variant Short Tandem Repeat motif count, either as given by its ALT MC value
+    from the variant FORMAT field, or as a number given in the ALT on the form
+    '<STR123>'.
+    """
+
+    alt_mc = None
+    if variant_obj["alternative"] == ".":
+        return alt_mc
+
+    for sample in variant_obj["samples"]:
+        if sample["genotype_call"] in ["./.", ".|", "0/0", "0|0"]:
+            continue
+        alt_mc = sample.get("alt_mc")
+    if alt_mc:
+        return alt_mc
+
+    alt_num = NUM.match(variant_obj["alternative"])
+    if alt_num:
+        alt_mc = int(alt_num)
+        return alt_mc
+
+    return None
 
 
 def download_str_variants(case_obj, variant_objs):
@@ -995,9 +1026,7 @@ def download_str_variants(case_obj, variant_objs):
         variant_line.append(
             variant.get("str_display_ru", variant.get("str_ru", ""))
         )  # Reference repeat unit
-        variant_line.append(
-            variant.get("alternative", "").replace("STR", "").replace("<", "").replace(">", "")
-        )  # Estimated size
+        variant_line.append(get_str_mc(variant) or ".")  # Estimated size
         variant_line.append(str(variant.get("str_ref", "")))  # Reference size
         variant_line.append(str(variant.get("str_status", "")))  # Status
         gt_cell = ""